[swig] Provide a way to force static bindings
The current logic in prepare_bindings.py makes it impossible to force it
to use the static bindings. You either have to pass the swig executable
or allow the script to try to find it itself. This is undesirable when
you just want to use the static bindings, for example when the version
of swig is too old.