[ScriptInterpreter] Move ownership into debugger (NFC)
This is part two of the change started in r359330. This patch moves the
ownership of the script interpreter from the command interpreter into
the debugger. I would've preferred to remove the lazy initialization,
however the fact that the scripting language is set after the debugger
is created makes that tricky. So for now this does exactly the same
thing as when it was under the command interpreter. The result is that
this patch is fully NFC.

Differential revision: https://reviews.llvm.org/D61211

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@359354 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 430fc972be3df0096a3b4df80ced9bfb9870b38b)