[CommandObject] Use GetDebugger() helper method (NFC)
In r359354 a GetDebugger() method was added to the CommandObject class,
so that we didn't have to go through the command interpreter to obtain
the script interpreter. This patch simplifies other call sites where
m_interpreter.GetDebugger() was used, and replaces them with a shorter
call to the new method.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@359373 91177308-0d34-0410-b5e6-96231b3b80d8