DWARF/NFC: Centralize DIERef conversions
apple and manual indexing code were creating a DIERef in a bunch of
places. Though the code itself is not much, it is also easy to simplify
by factoring out the DIERef creation. In HashedNameToDIE I create a
conversion operator from DIEInfo to DIERef, and in ManualDWARFIndex I
just create the DIERef in a global variable up-front.

This also reduces the diff in follow-up patches which change how DIERefs
are constructed.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@360669 91177308-0d34-0410-b5e6-96231b3b80d8