Simplify Triple::ppc64{,le} checks with Triple::isPPC64()
While here, update some ppc64le specific check to isPPC64(), if it
applies to big-endian as well, in the hope that it will ease the support
of big-endian if people are interested in this area. The big-endian
variant is used by at least FreeBSD, Gentoo Linux, Adélie Linux, and
Void Linux.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@360868 91177308-0d34-0410-b5e6-96231b3b80d8