[SwiftASTContext] Remove use of ClusteredBitVector::enumerateSetBits
The set bit enumeration API is being removed as part of
https://github.com/apple/swift/pull/24857.

This change replaces the enumerateSetBits call with a loop over all
the bits in the mask. This is potentially more expensive however it's
unlikely that the difference is noticeable given that we are already
building a mask of the same size byte-by-byte in this function.