repl: exit with failure if swiftCore.dll fails to load
Rather than print out an error message if the swiftCore.dll module fails
to load, simply exit with a status.  A subsequent change will change
this repl binary to a Win32 binary to avoid showing the console window,
so we will be unable to see the message in the first place.