Windows: define `ComputeSwiftDirectory` for Windows
Assume that the Windows layout is relatively similar to that on Linux.
`lldb.dll` is located in the `bin` directory rather than `lib` and the
`ComputeSupportExeDirectory` provides a `FileSpec` that is off by a
level.  Using `ComputeSupportExeDirectory` with `..\lib\swift` gives us
the actual location of the resource dir when lldb is installed into a
toolchain (as per the MSIs from
https://compnerd.visualstudio.com/windows-swift).