[lldb] Fix use-of-uninitialized-value in Driver
The driver passes by reference an uninitialized num_errors variable to
RunCommandInterpreter. This should be fine, as it's supposed to be
an output argument, but the reproducer instrumentation reads it in order
to record the value of all the arguments to the function.

This change fixes it by initializing num_errors to 0 before calling
RunCommandInterpreter.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@361444 91177308-0d34-0410-b5e6-96231b3b80d8