Make CompileUnit::GetSupportFiles return a const list
There's no reason for anyone to modify a list from outside of a symbol
file (as that would break a lot of invariants that symbol files depend
on).

Make the function return a const FileSpecList and fix up a couple of
places that were needlessly binding non-const references to the result
of this function.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@362069 91177308-0d34-0410-b5e6-96231b3b80d8