[LanguageRuntime] Hoist SwiftLanguageRuntime::IsSymbolARuntimeThunk
There's nothing swift-specific about thunks. There was a static function
in LanguageRuntime to determine if a symbol is a runtime thunk, but this just
creates a circular dependency between LanguageRuntime and SwiftLanguageRuntime.
I think if we hoist the one in SwiftLanguageRuntime into LanguageRuntime and
make it a virtual method, we should have a better, more general implementation.