Don't check the validity of newly contructed data buffers
A bunch of places were checking that DataBufferHeap::GetBytes returns a
non-null pointer right after constructing it. The only time when
GetBytes returns a null pointer is if it is empty (and I'm not sure that
even this is a good idea), but that is clearly not the case here, as the
buffer was constructed with a non-zero size just a couple of lines back.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@364754 91177308-0d34-0410-b5e6-96231b3b80d8