[Docs] Unify build instructions
The current build instructions are structured according the host
platform. Instead of having instructions on how to build with CMake
repeated for every platform, I unified them, using subsections if things
are different for between platforms. I also added the code signing
instructions, which were hidden in a text file in the repository.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@365081 91177308-0d34-0410-b5e6-96231b3b80d8