[Swift/Optional] Fix the formatter to print the right value.
There's a peculiarity in the way lldb represents `Int/UInt` types
in Swift. They don't have summaries, but values. This historical
choice forces us to switch between GetSummary() and GetValue()
in the formatters. Maybe we should revisit this choice because it
caused several issues over the last couple of years.

<rdar://problem/37835998>