[cmake] adjust NATIVE build for Swift
TableGen is a host tool and requires a native variant for every build.
While building as a part of llvm this is trivial and llvm handles it.
However, building standalone means that lldb has to handle this itself.
Add a NATIVE build variant to enable this.

This was landed upstream and needs SWIFT varaibles added accordingly.