[lldb] Tablegenify thread commands and fix completion bug for thread step-*
Beside turning the options into the new tablegen format, this patch
also fixes that a few commands had source file completions for the
"count" and "end-linenumber" arguments (which both accepted only
integers). Reason for that are that somehow we added a '1' instead
of our usual '0' value to the initial value for completion.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@366425 91177308-0d34-0410-b5e6-96231b3b80d8