Add an expectedFailure test for type finding.
When two .c files define a type of the same name, lldb
just picks one and uses it regardless of context.  That is
not correct.  When stopped in a frame in one of the .c files
that define this type, it should use that local definition.

This commit just adds a test that checks for the correct
behavior.  It is currently xfailed.


git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@366507 91177308-0d34-0410-b5e6-96231b3b80d8