Fix windows build after r366791
A side effect of this commit was that it exchanged the order of types
and compile units in the output of SymbolVendor::Dump. A couple of PDB
tests dependened on that to assert the links between the two.

While it wouldn't be too hard to update the tests, the change of
ordering was not something I intended to do with that patch, and is easy
to restore the original order, so I do just that.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@366798 91177308-0d34-0410-b5e6-96231b3b80d8