[lldb][CMake] Fix framework-enabled build detail for Xcode
If debugserver or any other framework tool gets built first, its post-build copy operation was using 'Resources' as the file name instead of the destination directory. It was not a problem with Ninja, because here the framework structure was alreaady created at configuration time. With this fix, both generators are happy.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@367005 91177308-0d34-0410-b5e6-96231b3b80d8