test: only pass `-sdk` on Darwin targets
On non-Darwin targets, the `-sdk` parameter may or may not contain the
sysroot which is needed to enable the use of `-sdk` to allow
cross-compilation.  This SDK contains the swift modules and the swift
runtime registrar.  Setting the SDKROOT to `/` on Linux requires that
the Swift SDK contents are installed to `/`.