PECOFF: Fix a "memset clearing an object of non-trivial type" warning
This time, the warning pointed to an actual problem, because the
coff_opt_header structure contained a std::vector. I guess this happened
to work because the all-zero state was a valid representation of an
empty vector, but its not a good idea to rely on that.

I remove the memset, and have the structure clear its members in the
constructor instead.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@367299 91177308-0d34-0410-b5e6-96231b3b80d8