[CompletionRequest] Remove unimplemented members.
Completion requests have two fields that are essentially unimplemented:
`m_match_start_point` and `m_max_return_elements`. This would've been
okay, if it wasn't for the fact that this caused a bunch of useless
parameters to be passed around. Occasionally there would be a comment or
assert saying that they are not supported. This patch removes them.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@367385 91177308-0d34-0410-b5e6-96231b3b80d8