[CMake] Always copy clang and swift resource directories when rebuilding
Currently the clang and swift resource dirs are only copied during configure
time. This could be a problem for iterative builds of lldb, where the swift
resource dir being built against could have changed. In order to make sure that
we have the most up-to-date clang and swift resource directories, we copy them
whenever we rebuild liblldb.