Module::SetArchitecture(): don't fail when there's no Swift context.
The recent change that made GetTypeSystemForLanguage return an
Expected changed the semantics of Module::SetArchitecture. It would
now fail when the Swift AST context is not found. This function is
just doing an additional tweak on the context when it exists, the
absence of Swift shouldn't make it fail.

Fixes Breakpoint/jitbp_elf.test

Of course, there shouldn't be any Swift specific code in this file in
the first place, but that's a battle for another day.