[dotest] Remove long running test "decorator" and re-enable tests.
Today I discovered the skipLongRunningTest decorator and to my surprise
all the tests were passing without the decorator. They don't seem to be
that expensive either, they take a few seconds but we have tests that
take much longer than that. As such I propose to remove the decorator
and enable them by default.

Differential revision: https://reviews.llvm.org/D66774

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@369995 91177308-0d34-0410-b5e6-96231b3b80d8