[test] Addres  TestConcurrentMany*.py flakiness on macOS
On "fast" macOS machines, the TestConcurrentMany*.py tests would fail
randomly with different numbers of breakpoints, watchpoints, etc. This
seems to be avoidable by giving the threads a little time to breath
after the passing the synchronization barrier. This is far from a
structural fix but it reduces the flakiness.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@370785 91177308-0d34-0410-b5e6-96231b3b80d8