[lldbrepl] Rewrite REPLTest after PExpectTest refactoring
PExpectTest was refactored upstream and none of the code shared with
REPLTest survived. Given the differences between the two types of tests,
I don't see the need for a common abstraction.

This rewrites REPLTest to make it functional again. Probably the biggest
change is that it no longer inherits from PExpectTest.