[lldb][NFC] Add dimensions to pexpect tests failing on sanitizer bot
I'm still unable to reproduce this locally, so let's see if this
is enough to make the bots happy. The missing dimension parameter
seems to be the biggest difference between these tests and
TestIOHandlerCompletion.py which runs just fine on the same bot,
so that's the motivation for trying this out.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@371077 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit 781e9606ac23ff83d8c7b4b10c3962fcb896134c)