[Reproducer] Add a `cont` to ModuleCXX.test
On more than one occasion I've found this test got stuck during replay
while waiting for a packet from debugserver when the debugger was in the
process of being destroyed. For some reason it's more prevalent on the
downstream Swift fork. Adding a cont mitigates the problem while I
investigate.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@371144 91177308-0d34-0410-b5e6-96231b3b80d8