[Reproducer] Disconnect when the replay server is out of packets.
This is a fix for the issue described in r371144.

> On more than one occasion I've found this test got stuck during replay
> while waiting for a packet from debugserver when the debugger was in
> the process of being destroyed.

When the replay server is out of packets we should just disconnect so
the debugger doesn't have to do any cleanup that it wouldn't do during
capture.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@371459 91177308-0d34-0410-b5e6-96231b3b80d8
(cherry picked from commit c15204b9f99b054858cfa9b22c6f86cf9b65df13)