[lldb][NFC] Remove ArgEntry::ref member
The StringRef should always be identical to the C string, so we
might as well just create the StringRef from the C-string. This
might be slightly slower until we implement the storage of ArgEntry
with a string instead of a std::unique_ptr<char[]>. Until then we
have to do the additional strlen on the C string to construct the
StringRef.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@371842 91177308-0d34-0410-b5e6-96231b3b80d8