[lldb] Make cursor index in CompletionRequest unsigned
The fact that index==-1 means "no arguments" is not obvious and only
used in one place from what I can tell. Also fixes several warnings
about using the cursor index as if it was a size_t when comparing.

Not fully NFC as we now also correctly update the partial argument list
when injecting the fake empty argument in the CompletionRequest
constructor.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@372566 91177308-0d34-0410-b5e6-96231b3b80d8