[lldb] Remove redundant argument lists in CompletionRequest
We currently have two lists in the CompletionRequest that we
inherited from the old API: The complete list of arguments ignoring
where the user requested completion and the list of arguments that
stops at the cursor. Having two lists of arguments is confusing
and can lead to subtle errors, so let's remove the complete list
until we actually need it.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@372692 91177308-0d34-0410-b5e6-96231b3b80d8