SystemInitializer: Use Targets.def to selectively initialize ABI plugins
This avoids having to define additional macros in the cmake file, and
and also makes the logic in the cpp files more compact. It is also
easily extendible to other plugin types (instruction emulation?) that
should only be initialized if the corresponding llvm target is built.

Thanks to Ilya Birukov for pointing me to this file.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@372952 91177308-0d34-0410-b5e6-96231b3b80d8