[CMake] Add option to not copy over the Swift resources
The Swift headers are needed for development and testing, but should not
be present in the LLDB.framework when it's installed. I couldn't figure
out a good way to prevent just these files from getting installed, so
instead this adds a configuration option to prevent the headers from
getting copied in the first place.