[Reproducer] Always use absolute paths for capture & replay.
The VFS requires files to be have absolute paths. The file collector
makes paths relative to the reproducer root. If the root is a relative
path, this would trigger an assert in the VFS. This patch ensures that
we always make the given path absolute.

Thank you Ted Woodward for pointing this out!

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@373102 91177308-0d34-0410-b5e6-96231b3b80d8