Fix sign extension handling in DumpEnumValue
When an enumerator has an unsigned type and its high bit set, the
code introduced in r374067 would fail to match it due to a sign
extension snafu. This commit fixes this aspec of the code and should
fix the bots.

I think it's not a complete fix though, I'll add more test coverage
and additional tweaks in a follow-up commit.

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@374095 91177308-0d34-0410-b5e6-96231b3b80d8